# docs/swagger.yaml
basePath: /
definitions:
  api.Accept:
    properties:
      flag:
        type: boolean
    type: object
  api.Audit:
    properties:
      limit:
        $ref: '#/definitions/api.Limit'
    type: object
  api.Destination:
    properties:
      address:
        type: string
      invert:
        type: string
    type: object
  api.Drop:
    properties:
      flag:
        type: boolean
    type: object
  api.ForwardPort:
    properties:
      port:
        type: string
      protocol:
        type: string
      toaddr:
        type: string
      toport:
        type: string
    type: object
  api.IcmpBlock:
    properties:
      name:
        type: string
    type: object
  api.IcmpType:
    properties:
      name:
        type: string
    type: object
...
  api.Limit:
    properties:
      value:
        type: string
    type: object
  api.Log:
    properties:
      level:
        type: string
      limit:
        $ref: '#/definitions/api.Limit'
      prefix:
        type: string
    type: object
  api.Mark:
    properties:
      limit:
        $ref: '#/definitions/api.Limit'
      set:
        type: string
    type: object
  api.Port:
    properties:
      port:
        type: string
      protocol:
        type: string
    type: object
  api.Protocol:
    properties:
      value:
        type: string
    type: object
...
  api.Reject:
    properties:
      flag:
        type: boolean
    type: object
  api.Rule:
    properties:
      accept:
        $ref: '#/definitions/api.Accept'
      audit:
        $ref: '#/definitions/api.Audit'
      destination:
        $ref: '#/definitions/api.Destination'
      drop:
        $ref: '#/definitions/api.Drop'
      family:
        type: string
      forwardport:
        $ref: '#/definitions/api.ForwardPort'
      icmpblock:
        $ref: '#/definitions/api.IcmpBlock'
      icmptype:
        $ref: '#/definitions/api.IcmpType'
      limit:
        $ref: '#/definitions/api.Limit'
      log:
        $ref: '#/definitions/api.Log'
      mark:
        $ref: '#/definitions/api.Mark'
      port:
        $ref: '#/definitions/api.Port'
      protocol:
        $ref: '#/definitions/api.Protocol'
      reject:
        $ref: '#/definitions/api.Reject'
      service:
        items:
          type: string
        type: array
      source:
        $ref: '#/definitions/api.Source'
      value:
        $ref: '#/definitions/api.Value'
    type: object
...
  api.Source:
    properties:
      address:
        type: string
      invert:
        type: string
      ipset:
        type: string
      mac:
        type: string
    type: object
...
  api.Value:
    properties:
      value:
        type: string
    type: object
  query.AllocateRoleQuery:
    properties:
      role_ids:
        items:
          type: integer
        type: array
      user_id:
        type: integer
    required:
    - role_ids
    - user_id
    type: object
...
  query.BatchPortQuery:
    properties:
      delay:
        type: integer
      ports:
        items:
          $ref: '#/definitions/query.PortQuery'
        type: array
    type: object
...
  query.BatchServiceQuery:
    properties:
      delay:
        type: integer
      services:
        items:
          $ref: '#/definitions/query.ServiceQuery'
        type: array
    type: object
  query.BatchSettingQuery:
    properties:
      delay:
        type: integer
      hosts:
        items:
          type: string
        type: array
    required:
    - hosts
    type: object
...
  query.PortQuery:
    properties:
      ip:
        type: string
      port:
        $ref: '#/definitions/api.Port'
      timeout:
        type: integer
      zone:
        type: string
    required:
    - ip
    - port
    type: object
  query.Query:
    properties:
      forward:
        $ref: '#/definitions/api.ForwardPort'
      ip:
        type: string
      port:
        $ref: '#/definitions/api.Port'
      rich:
        $ref: '#/definitions/api.Rule'
      service:
        type: string
      timeout:
        type: integer
      zone:
        type: string
    required:
    - ip
    type: object
...
  query.ServiceQuery:
    properties:
      ip:
        type: string
      service:
        type: string
      timeout:
        type: integer
      zone:
        type: string
    required:
    - ip
    - service
    type: object
...
  query.UserEditQuery:
    properties:
      id:
        type: integer
      password:
        type: string
      role_ids:
        items:
          type: integer
        type: array
      username:
        type: string
    required:
    - username
    type: object
  query.UserQuery:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
host: localhost:2952
info:
  contact: {}
  description: Uranus, distrubed firewall gateway.
  title: Uranus API
  version: 0.0.9
paths:
...
  /fw/v1/service:
    delete:
      consumes:
      - application/json
      description: Remove a service rule on firewalld runtime.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.ServiceQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Remove a service rule on firewalld runtime.
      tags:
      - firewalld service
    get:
      consumes:
      - application/json
      description: List all service rule on firewalld runtime.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.ServiceQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List available service type on firewalld runtime.
      tags:
      - firewalld service
    put:
      consumes:
      - application/json
      description: Add a service rule on firewalld runtime.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.ServiceQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add a service rule on firewalld runtime.
      tags:
      - firewalld service
...
  /fw/v2/service:
    delete:
      consumes:
      - application/json
      description: Remove a service rule on firewalld premanent.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.ServiceQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Remove a service rule on firewalld premanent.
      tags:
      - firewalld service
    get:
      consumes:
      - application/json
      description: List all service rule on firewalld permanent.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.Query'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List available service type on firewalld permanent.
      tags:
      - firewalld service
    put:
      consumes:
      - application/json
      description: Permanently add service to list of services used in zone.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.ServiceQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Permanently add service to list of services used in zone.
      tags:
      - firewalld service
...
  /fw/v3/service:
    put:
      consumes:
      - application/json
      description: Add a service rule on firewalld runtime with delay timer.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.BatchServiceQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add a service rule on firewalld runtime with delay timer.
      tags:
      - firewalld service
  /fw/v3/setting/reload/runtime:
    put:
      consumes:
      - application/json
      description: Reload firewalld runtime with delay timer.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.BatchSettingQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Reload firewalld runtime with delay timer.
      tags:
      - firewalld setting
  /fw/v3/setting/sdzone:
    post:
      consumes:
      - application/json
      description: Set default zone on firewalld with delay timer, this is a runtime
        and permanent change.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.BatchPortQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Set default zone on firewalld with delay timer, this is a runtime and
        permanent change.
      tags:
      - firewalld setting
...
  /security/users:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      description: Delete a user.
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete a user.
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Return all users。
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      - description: sort
        in: query
        name: sort
        type: string
      - description: sort
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Return all users.
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a user。
      parameters:
      - description: Input parameter
        in: body
        name: userinput
        required: true
        schema:
          $ref: '#/definitions/query.UserEditQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: update user information.
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Create a user。
      parameters:
      - description: Input parameter
        in: body
        name: userinput
        required: true
        schema:
          $ref: '#/definitions/query.UserQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create a user.
      tags:
      - Users
  /security/users/allocate:
    post:
      consumes:
      - application/json
      description: Assign roles to users。
      parameters:
      - description: Input parameter
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/query.AllocateRoleQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Assign roles to users.
      tags:
      - Users
  /sso/signin:
    post:
      consumes:
      - application/json
      description: login.
      parameters:
      - description: signup body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.UserQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: login.
      tags:
      - SSO
  /sso/signup:
    post:
      consumes:
      - application/json
      description: register.
      parameters:
      - description: user body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.UserQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: register.
      tags:
      - SSO
schemes:
- http
securityDefinitions:
  Bearer:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"